+++ src/itemRadioGroup/index.tsx
@@ -12,6 +12,7 @@ export interface ItemRadioGroupProps {
   readonly status?: ItemRadioStatus
   readonly withSeparators?: boolean
   readonly withChevrons?: boolean
+  readonly ariaLabelledBy?: string
 }
 
 interface ItemRadioGroupState {
@@ -37,9 +38,22 @@ class ItemRadioGroup extends PureComponent<ItemRadioGroupProps, ItemRadioGroupSt
   }
 
   render() {
-    const { children, status, name, className, withSeparators, withChevrons } = this.props
+    const {
+      children,
+      status,
+      name,
+      className,
+      withSeparators,
+      withChevrons,
+      ariaLabelledBy,
+    } = this.props
     return (
-      <ItemsList withSeparators={withSeparators} className={className} role="radiogroup">
+      <ItemsList
+        withSeparators={withSeparators}
+        className={className}
+        role="radiogroup"
+        aria-labelledby={ariaLabelledBy}
+      >
         {children.map(item => {
           const itemProps: Partial<ItemRadioProps> = item.props
           const checked = this.state.value === itemProps.value